requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
requests
msal
cryptography
pyOpenSSL
python-multipart
pypdf
reportlab
num2words
